requirements.txt
google-genai
websockets
pypdf
